put methods in config.utils_method
