+++ apps/sales/views.py
@@ -16,8 +16,7 @@ from django.shortcuts import get_object_or_404
 from rest_framework.filters import OrderingFilter
 from django_filters.rest_framework import DjangoFilterBackend  # type: ignore
 from .filters import SaleOrderFilter, SaleInvoiceOrdersFilter, SaleReturnOrdersFilter
-from apps.sales.utils.document_generator_methods import save_sales_order_pdf_to_media, extract_product_data, convert_amount_to_words
-from config.utils_methods import format_phone_number,send_pdf_via_email, send_whatsapp_message_via_wati, get_related_data, update_multi_instances, validate_input_pk, delete_multi_instance, generic_data_creation, get_object_or_none, list_all_objects, create_instance, update_instance, build_response, validate_multiple_data, validate_order_type, validate_payload_data, validate_put_method_data
+from config.utils_methods import convert_amount_to_words, extract_product_data, save_sales_order_pdf_to_media, format_phone_number,send_pdf_via_email, send_whatsapp_message_via_wati, get_related_data, update_multi_instances, validate_input_pk, delete_multi_instance, generic_data_creation, get_object_or_none, list_all_objects, create_instance, update_instance, build_response, validate_multiple_data, validate_order_type, validate_payload_data, validate_put_method_data
 
 # Set up basic configuration for logging
 logging.basicConfig(level=logging.INFO,
